refactor: move web/email index files one level down
